Give SCons better default flags and `-ansi` flag
I messed around with the compiler flags, the best trade-off I was able to
get was from using the `-Os -flto` flags (mingw), the current default will
just be `-O2`.

`-O3` doesn't seem to do much other than increase the file size, but samples_c
might not be the best benchmark tool.

The `-ansi` flag will give errors if newer compiler features are used in a
backwards incompatible way.  This flag is not set for sample code, only
libtcod.